requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
pyodbc==4.0.32
PyJWT==2.1.0
python-dotenv==0.19.0
werkzeug==2.0.1
